requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
django-cors-headers
